Move Zyte related code to own module
