+++ spidermon/contrib/monitors/mixins/job.py
@@ -1,9 +1,1 @@
-from spidermon.exceptions import NotConfigured
-
-
-class JobMonitorMixin:
-    @property
-    def job(self):
-        if not self.data.job:
-            raise NotConfigured("Job not available!")
-        return self.data.job
+from spidermon.contrib.zyte.monitors.mixins.job import JobMonitorMixin  # noqa
