+++ spidermon/contrib/zyte/monitors/mixins/job.py
@@ -1,0 +1,9 @@
+from spidermon.exceptions import NotConfigured
+
+
+class JobMonitorMixin:
+    @property
+    def job(self):
+        if not self.data.job:
+            raise NotConfigured("Job not available!")
+        return self.data.job
